+++ components/todoCard/filters/Filters.module.scss
@@ -1,0 +1,32 @@
+.filtersWrapper{
+    background-color: #d5d5d5;
+    width: 100%;
+    height: 40px;
+    box-shadow: 0px 1px 8px #8d8d8d;
+    display: flex;
+    flex-direction: row;
+    justify-content: space-between;
+    padding: 0px 15px;
+    align-items: center;
+    .filterButtonWrapper{
+        .filterButton{
+            margin: 4px;
+            cursor: pointer;
+        }
+        .active{
+            color: #000;
+            font-weight: 600;
+        }
+    }
+    .clearButton{
+        border: none;
+        outline: none;
+        width: 140px;
+        height: 25px;
+        background-color: #ededed;
+        cursor: pointer;
+    }
+    .clearButton:hover{
+        background-color: #bdbdbd;
+    }
+}
