+++ components/todoCard/taskCellS/TaskCells.module.scss
@@ -1,0 +1,160 @@
+.cardWrapper{
+    height: 70vh;
+    height: 70vh;
+    width: 50vw;
+    background-color: #ededed;
+    border-top: 20px solid #646464;
+    border-radius: 2px;
+    box-shadow: 0px 1px 8px #8d8d8d;
+    padding-bottom: 50px;
+    .filtersWrapper{
+        background-color: #d5d5d5;
+        width: 100%;
+        height: 40px;
+        box-shadow: 0px 1px 8px #8d8d8d;
+        display: flex;
+        flex-direction: row;
+        justify-content: space-between;
+        padding: 0px 15px;
+        align-items: center;
+        .filterButtonWrapper{
+            .filterButton{
+                margin: 4px;
+                cursor: pointer;
+            }
+            .active{
+                color: #000;
+                font-weight: 600;
+            }
+        }
+        .clearButton{
+            border: none;
+            outline: none;
+            width: 140px;
+            height: 25px;
+            background-color: #ededed;
+            cursor: pointer;
+        }
+        .clearButton:hover{
+            background-color: #bdbdbd;
+        }
+    }
+}
+.inputSection{
+    height: 50px;
+    background-color: #d5d5d5;
+    border-bottom: 1px solid #adadad;
+    display: flex;
+    flex-direction: row;
+    .customInput{
+        background-color: transparent;
+        // background-color: red;
+        width: 90%;
+        height: 100%;
+        border: none;
+        outline: none;
+        padding: 20px;
+        font-size: 20px;
+    }
+    .customInput::placeholder{
+        font-style: italic;
+    }
+    .buttonWrapper{
+        // background-color: yellow;
+        flex: 1;
+        padding: 5px
+    }
+    .addButton{
+        border: none;
+        outline: none;
+        cursor: pointer;
+        // background-color: yellow;
+        width: 100%;
+        height: 100%;
+        border-radius: 20px;
+        .buttonText{
+            font-size: 20px;
+            color: #646464;
+            font-weight: 600;
+        }
+    }
+    .addButton:hover{
+        background-color: rgb(255, 255, 215);
+    }
+    
+}
+.noItemExist{
+    font-size: 20px;
+    color: #646464;
+    font-weight: 600;
+    text-align: center;
+    padding: 20px;
+}
+
+
+
+.listCointainer{
+    overflow-y: scroll;
+    height: 100%;
+    .rowContainer{
+        // background-color: red;
+        height: 50px;
+        max-height: 80px;
+        display: flex;
+        flex-direction: row;
+        align-items: center;
+        border-bottom: 1px solid #adadad;
+        .actionWrapper{
+            // background-color: green;
+            height: 100%;
+            min-width: 40px;
+            display: flex;
+            justify-content: center;
+            align-items: center;
+            cursor: pointer;
+            .checkbox{
+                width: 20px;
+                height: 20px;
+                cursor: pointer;
+            }
+        }
+        .removeItem{
+            display: none;
+        }
+        .taskMessageWrapper{
+            padding-left: 10px;
+            flex: 1;
+            // background-color: red;
+            cursor: pointer;
+            height: 100%;
+            display: flex;
+            align-items: center;
+            .taskMessage{
+                font-size: 20px;
+            }
+            .updateTaskInputSection{
+                width: 100%;
+                height: 100%;
+                // background-color: yellow;
+                .updateTaskInput{
+                    width: 100%;
+                    height: 100%;
+                    background-color: inherit;
+                    font-size: 20px;
+                    outline: none;
+                    border: 1px solid #646464;
+                }
+            }
+        }
+    }
+    .rowContainer:hover{
+        .removeItem{
+            display: inherit;
+            padding: 10px;
+            .actionText{
+                font-weight: 600;
+            }
+        }
+    }
+}
+
